+++ src/onegov.election_day/onegov/election_day/forms/vote.py
@@ -1,9 +1,11 @@
 from datetime import date
+from onegov.ballot import VoteCollection
+from onegov.core import utils
 from onegov.election_day import _
 from onegov.form import Form
 from wtforms import RadioField, StringField
 from wtforms.fields.html5 import DateField
-from wtforms.validators import InputRequired
+from wtforms.validators import InputRequired, ValidationError
 
 
 class VoteForm(Form):
@@ -26,6 +28,12 @@ class VoteForm(Form):
         ('federation', _("Federal Vote")),
         ('canton', _("Cantonal Vote")),
     ], validators=[InputRequired()])
+
+    def validate_vote(self, field):
+        votes = VoteCollection(self.request.app.session())
+
+        if votes.by_id(utils.normalize_for_url(field.data)):
+            raise ValidationError(_("A vote with this title exists already"))
 
     def update_model(self, model):
         model.title = self.vote.data